Implement nested commenting functionality.
Notepad++ has a very interesting XML plugin which deals with
nested XML comments in a very simple way:

* Each time a selection gets commented, any existing nested comments are
replaced with '<!{1}**' and '**{1}>'. These characters are no longer
special and thus allowed inside of a comment.
* Uncommenting will remove any comment tags and place new comment tags
everywhere the placeholders '<!{1}**' and '**>' can be found.
* If commenting a block of text which already has any of those
placeholders, the numbers in brackets are increased.
If uncommenting, the numbers are decreased.
This makes commenting/uncommenting a lossless operation, which is a
large improvement compared to other editors where the nested comments
are simply removed.

By default the following mappings are implemented for visual modes:
* <Localleader>c
* <Localleader>u