Fixes normal insertion of right angle bracket
Since 781eccae there were problems inserting a right angle bracket in
some situations causing the inserted character to be one step to the
right of where it should've been. The reason being that "normal! l" was
run in every situation. This commit changes it so that it's only run in
the cases where it's needed and the normal case is that we just do
"normal! a>".

Fixes #44